extract DatabaseConfig from SQLStore
